One hot

+++ task_1.py
@@ -9,6 +9,6 @@ data = pd.DataFrame({'whoAmI':lst})
 new_data = pd.DataFrame({'robot' : [],
                          'human' : []})
 
-new_data['human'] = data['whoAmI'] == 'human'
-new_data['robot'] = data['whoAmI'] == 'robot'
+new_data['human'] = (data['whoAmI'] == 'human').astype(int)
+new_data['robot'] = (data['whoAmI'] == 'robot').astype(int)
 print(new_data)
